Adding the test client.
The call backs work but there is no proper way to get he data as of now. This
means that more cffi stuff needs to be done.

In the interest of time we will move to another approach i.e simply hack the
print stream in c interface and make a steraming server